+++ reference/src/main/java/com/ebito/reference/repository/OperationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OperationRepository extends JpaRepository<Operation, Integer> {
+public interface OperationRepository extends JpaRepository<Operation, Long> {
     List<Operation> findAllByClientId(long clientId);
 }
